+++ active-record/src/test/java/com/iluwatar/activeobject/AppTest.java
@@ -1,6 +1,5 @@
 package com.iluwatar.activeobject;
 
-import com.iluwatar.activeobject.App;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
